scraping_threaded/requirements.txt
annotated-types==0.5.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
ciso8601==2.2.0
click==8.1.6
dnspython==2.4.2
idna==3.4
lxml==4.9.3
mediacloud==4.0.1
pydantic==2.1.1
pydantic_core==2.4.0
pymongo==4.4.1
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.4.1
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.4
wayback-news-search==1.0.1
